requirements.txt
gunicorn
flask
tensorflow==2.3.0
pillow
numpy
protobuf==3.20.*
